♻️ refactor: SRP, CQRS 적용
